Fixed additional issue where type was inferred even though it was a typedef
